Adding user access and other features

+++ base/views.py
@@ -3,17 +3,22 @@ from http.client import HTTPResponse
 from os import execv
 from pickle import GLOBAL
 from django.shortcuts import render , redirect
-from django.http import HttpResponse
+from django.http import HttpRequest, HttpResponse
 import sqlite3
 from django.contrib.auth import login , authenticate , logout
 from django.contrib import messages
 import time
 
+from datetime import datetime
+import pytz
+
 def visualRepresent(request) :
+    tz_kol = pytz.timezone('Asia/Calcutta') 
+    datetime_kol = datetime.now(tz_kol)
 
     try:
         energy = []
-        time = []
+        times = []
         conn = sqlite3.connect("Data.db")
         c = conn.cursor()
         c.execute("SELECT * FROM datas ;")
@@ -23,12 +28,12 @@ def visualRepresent(request) :
         for i in values : 
             
             energy.append(float(i[2])/1000)
-            time.append(i[3])
+            times.append(i[3])
             x = x + 1
         
     except:
         pass
-    return render (request , 'visualization.html' , {"energy" : energy , "time" : time})
+    return render (request , 'visualization.html' , {"energy" : energy , "times" : times})
 
 
 def updateData(request) :
@@ -49,9 +54,13 @@ def updateData(request) :
         current = float(request.POST.get('current'))
         print(voltage*current)
         energy = voltage*current
+
+        tz_kol = pytz.timezone('Asia/Calcutta') 
+        datetime_kol = datetime.now(tz_kol)
         
-        named_tuple = time.localtime()
-        time_INT = int(time.strftime("%H%M%S", named_tuple))
+        time_INT = int(datetime_kol.strftime("%H%M%S"))
+        # named_tuple = time.localtime()
+        # time_INT = int(time.strftime("%H%M%S", named_tuple))
         c.execute(f"""INSERT INTO datas VALUES({ current },{ voltage },{ energy },{time_INT})""")
         conn.commit()
         conn.close()
@@ -79,16 +88,23 @@ def home(request):
 
 
 def deleteDB(request):
-    try: 
-        conn = sqlite3.connect("Data.db")
-        c = conn.cursor()
-        c.execute("DROP TABLE datas;")
-        conn.commit()
-        conn.close()
-        return HttpResponse("Deleted")
-    except :
-        print("Table Doesn't Exist")
-        return HttpResponse()
+    if request.method == "POST" :
+            username = request.POST['username']
+            password = request.POST['password']
+            user = authenticate(request, username=username, password=password)
+            if user is not None:
+                login(request, user)
+                try :
+                    conn = sqlite3.connect("Data.db")
+                    c = conn.cursor()
+                    c.execute("DROP TABLE datas;")
+                    conn.commit()
+                    conn.close()
+                    return redirect(home)
+                except:
+                    return redirect(home)
+            else:
+                return redirect('delete')
 
-def login(request):
-    return 
+    else : 
+        return render(request , 'deleteDBform.html')
